requirements.txt
pdf2image
pillow
openai
python-dotenv
tqdm



